refactor: 티켓 예매하기 티켓으로 이동
