+++ android/festago/app/src/main/java/com/festago/festago/data/service/ReservationRetrofitService.kt
@@ -1,12 +1,8 @@
 package com.festago.festago.data.service
 
 import com.festago.festago.data.dto.ReservationTicketsResponse
-import com.festago.festago.data.dto.ReservedTicketRequest
-import com.festago.festago.data.dto.ReservedTicketResponse
 import retrofit2.Response
-import retrofit2.http.Body
 import retrofit2.http.GET
-import retrofit2.http.POST
 import retrofit2.http.Path
 
 interface ReservationRetrofitService {
@@ -14,9 +10,4 @@ interface ReservationRetrofitService {
     suspend fun getTickets(
         @Path("stageId") stageId: Int,
     ): Response<ReservationTicketsResponse>
-
-    @POST("/member-tickets")
-    suspend fun postReserveTicket(
-        @Body ticketId: ReservedTicketRequest,
-    ): Response<ReservedTicketResponse>
 }
