+++ android/festago/app/src/main/java/com/festago/festago/data/service/TicketRetrofitService.kt
@@ -2,8 +2,11 @@ package com.festago.festago.data.service
 
 import com.festago.festago.data.dto.MemberTicketResponse
 import com.festago.festago.data.dto.MemberTicketsResponse
+import com.festago.festago.data.dto.ReservedTicketRequest
+import com.festago.festago.data.dto.ReservedTicketResponse
 import com.festago.festago.data.dto.TicketCodeDto
 import retrofit2.Response
+import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
 import retrofit2.http.Path
@@ -27,4 +30,9 @@ interface TicketRetrofitService {
     suspend fun getHistoryTickets(
         @Query("size") size: Int,
     ): Response<MemberTicketsResponse>
+
+    @POST("/member-tickets")
+    suspend fun postReserveTicket(
+        @Body ticketId: ReservedTicketRequest,
+    ): Response<ReservedTicketResponse>
 }
