+++ android/festago/app/src/main/java/com/festago/festago/domain/repository/ReservationRepository.kt
@@ -1,9 +1,7 @@
 package com.festago.festago.domain.repository
 
 import com.festago.festago.domain.model.ReservationTicket
-import com.festago.festago.domain.model.ReservedTicket
 
 interface ReservationRepository {
     suspend fun loadTicketTypes(stageId: Int): Result<List<ReservationTicket>>
-    suspend fun reserveTicket(ticketId: Int): Result<ReservedTicket>
 }
