+++ android/festago/app/src/main/java/com/festago/festago/domain/repository/TicketRepository.kt
@@ -1,5 +1,6 @@
 package com.festago.festago.domain.repository
 
+import com.festago.festago.domain.model.ReservedTicket
 import com.festago.festago.domain.model.Ticket
 import com.festago.festago.domain.model.TicketCode
 
@@ -8,4 +9,5 @@ interface TicketRepository {
     suspend fun loadCurrentTickets(): Result<List<Ticket>>
     suspend fun loadTicketCode(ticketId: Long): Result<TicketCode>
     suspend fun loadHistoryTickets(size: Int): Result<List<Ticket>>
+    suspend fun reserveTicket(ticketId: Int): Result<ReservedTicket>
 }
